- made the ResonanceMolSupplier really lazy, i.e. resonance structure   enumeration is only carried out when the user asks for a structure   or when the user explicitly request that calling the enumerate()   member function. This makes object creation fast and enables   calling getNumConjGrps(), getBondConjGrpIdx() and getAtomConjGrpIdx()   member function without incurring in the cost of necessarily   enumerating resonance structures - fixed some int/unsigned int inconsistencies - now bonds and atoms with do not belong to conjugated groups   get a -1 index - added a few Python wrappers - added a few tests - removed unused variables from Code/GraphMol/ForceFieldHelpers/MMFF/Builder.cpp - fixed ResonanceMolSupplier Java wrappers which were really broken

+++ Code/RDGeneral/RDThreads.h
@@ -20,7 +20,7 @@ namespace RDKit{
       return static_cast<unsigned int>(target);
     }
     unsigned int res=boost::thread::hardware_concurrency();
-    if(res>-target){
+    if(static_cast<int>(res) > -target){
       return res+target;
     } else {
       return 1;
